Override ActiveSupport::TestCase by default
This is a big change that simplifies the approach to how minitest is integrated.
Use ActiveSupport::TestCase instead of MiniTest::Rails::ActiveSupport::TestCase.
Instead of maintaining separate TestCase objects, overwrite the standard objects.
Focus only on on what is needed to enable minitest and the spec DSL.
Remove the MiniTest::Rails:: prefix from the generated tests.