Drop 7.1 files / identifiers from the master branch
Best I can tell, the intent is to keep ONE version on master or
versioned branch but for some reason 7.1 added a 7.2.beta and kept it
around for a bit. I think this brings it back to the original intent.

Long term: nuke the gemfiles directory. Remove gemfiles from the CI
matrix. Unify.